This are the example of creating the databased systems
in this example we are looking at the 1: we are looking at the department of electrical, electronics and computer engineering, we have to number of undergraduate qualifications that students can take, a student can enroll for one qualification at a time. so, we create the databased management system that can be used by the department.